requirements.txt
beautifulsoup4==4.6.0
bs4==0.0.1
Flask==1.0.3
gunicorn==19.7.1
Janome==0.3.6
Jinja2==2.10
mecab-python3==0.7
mysql-connector-python==8.0.17
nltk==3.4.4
numpy==1.15.4
urllib3==1.22